refactor: fixture 도메인 별로 분리 & 의존성 주입 순서 일치
