+++ src/test/java/org/greedy/ddarahang/common/fixture/RegionFixture.java
@@ -1,0 +1,13 @@
+package org.greedy.ddarahang.common.fixture;
+
+import org.greedy.ddarahang.db.country.Country;
+import org.greedy.ddarahang.db.region.Region;
+
+public class RegionFixture {
+    public static Region getMockRegion(Country country) {
+        return Region.builder()
+                .name("서울")
+                .country(country)
+                .build();
+    }
+}
